+++ src/shared/main/api/getStampStatus.ts
